[fix] (infrastructure/logic) :truck: fix dataset class path due to rename folder name (#85)
