+++ src/titanic_analysis/infrastructure/logic/analysis/analyzer.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, ClassVar, TypeVar
 
-from titanic_analysis.domain.dataset.dataset import Dataset
+from titanic_analysis.domain.dataset.sklearn_dataset import Dataset
 from titanic_analysis.infrastructure.io.utils import DisplayUtility
 
 if TYPE_CHECKING:
